Some improvements

+++ C/bpm_phase_beats.c
@@ -1,4 +1,4 @@
-void bpm_phase_beats(float32_t bpm, float32_t phase, uint32_t onsets_size, uint32_t fs)
+void bpm_phase_beats(float32_t bpm, float32_t phase, int8_t onsets_size, int8_t fs)
 {
 	float32_t i = phase;
 	float32_t beats_period = 1/(bpm/60);
